move proto tests to `proto_reader` module
