--- src/common/src/array/column_proto_readers.rs
@@ -1,203 +1,0 @@
-// Copyright 2023 RisingWave Labs
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use std::io::{Cursor, Read};
-
-use anyhow::anyhow;
-use byteorder::{BigEndian, ReadBytesExt};
-use paste::paste;
-use risingwave_pb::data::PbArray;
-
-use crate::array::value_reader::{PrimitiveValueReader, VarSizedValueReader};
-use crate::array::{
-    Array, ArrayBuilder, ArrayImpl, ArrayResult, BoolArray, DateArrayBuilder, IntervalArrayBuilder,
-    PrimitiveArrayBuilder, PrimitiveArrayItemType, TimeArrayBuilder, TimestampArrayBuilder,
-};
-use crate::buffer::Bitmap;
-use crate::types::{Date, Interval, Time, Timestamp};
-
-// TODO: Use techniques like apache arrow flight RPC to eliminate deserialization.
-// https://arrow.apache.org/docs/format/Flight.html
-
-pub fn read_numeric_array<T: PrimitiveArrayItemType, R: PrimitiveValueReader<T>>(
-    array: &PbArray,
-    cardinality: usize,
-) -> ArrayResult<ArrayImpl> {
-    ensure!(
-        array.get_values().len() == 1,
-        "Must have only 1 buffer in a numeric array"
-    );
-
-    let buf = array.get_values()[0].get_body().as_slice();
-
-    let mut builder = PrimitiveArrayBuilder::<T>::new(cardinality);
-    let bitmap: Bitmap = array.get_null_bitmap()?.into();
-    let mut cursor = Cursor::new(buf);
-    for not_null in bitmap.iter() {
-        if not_null {
-            let v = R::read(&mut cursor)?;
-            builder.append(Some(v));
-        } else {
-            builder.append(None);
-        }
-    }
-    let arr = builder.finish();
-    ensure_eq!(arr.len(), cardinality);
-
-    Ok(arr.into())
-}
-
-pub fn read_bool_array(array: &PbArray, cardinality: usize) -> ArrayResult<ArrayImpl> {
-    ensure!(
-        array.get_values().len() == 1,
-        "Must have only 1 buffer in a bool array"
-    );
-
-    let data = (&array.get_values()[0]).into();
-    let bitmap: Bitmap = array.get_null_bitmap()?.into();
-
-    let arr = BoolArray::new(data, bitmap);
-    ensure_eq!(arr.len(), cardinality);
-
-    Ok(arr.into())
-}
-
-fn read_date(cursor: &mut Cursor<&[u8]>) -> ArrayResult<Date> {
-    match cursor.read_i32::<BigEndian>() {
-        Ok(days) => Date::with_days(days).map_err(|e| anyhow!(e).into()),
-        Err(e) => bail!("Failed to read i32 from Date buffer: {}", e),
-    }
-}
-
-fn read_time(cursor: &mut Cursor<&[u8]>) -> ArrayResult<Time> {
-    match cursor.read_u64::<BigEndian>() {
-        Ok(t) => Time::with_nano(t).map_err(|e| anyhow!(e).into()),
-        Err(e) => bail!("Failed to read i64 from NaiveTime buffer: {}", e),
-    }
-}
-
-fn read_timestamp(cursor: &mut Cursor<&[u8]>) -> ArrayResult<Timestamp> {
-    cursor
-        .read_i64::<BigEndian>()
-        .map_err(|e| anyhow!("Failed to read i64 from Timestamp buffer: {}", e))
-        .and_then(|t| Timestamp::with_macros(t).map_err(|e| anyhow!("{}", e)))
-        .map_err(Into::into)
-}
-
-pub fn read_interval(cursor: &mut Cursor<&[u8]>) -> ArrayResult<Interval> {
-    let mut read = || {
-        let months = cursor.read_i32::<BigEndian>()?;
-        let days = cursor.read_i32::<BigEndian>()?;
-        let usecs = cursor.read_i64::<BigEndian>()?;
-
-        Ok::<_, std::io::Error>(Interval::from_month_day_usec(months, days, usecs))
-    };
-
-    match read() {
-        Ok(iu) => Ok(iu),
-        Err(e) => bail!("Failed to read Interval from buffer: {}", e),
-    }
-}
-
-macro_rules! read_one_value_array {
-    ($({ $type:ident, $builder:ty }),*) => {
-        paste! {
-            $(
-            pub fn [<read_ $type:snake _array>](array: &PbArray, cardinality: usize) -> ArrayResult<ArrayImpl> {
-                ensure!(
-                    array.get_values().len() == 1,
-                    "Must have only 1 buffer in a {} array", stringify!($type)
-                );
-
-                let buf = array.get_values()[0].get_body().as_slice();
-
-                let mut builder = $builder::new(cardinality);
-                let bitmap: Bitmap = array.get_null_bitmap()?.into();
-                let mut cursor = Cursor::new(buf);
-                for not_null in bitmap.iter() {
-                    if not_null {
-                        builder.append(Some([<read_ $type:snake>](&mut cursor)?));
-                    } else {
-                        builder.append(None);
-                    }
-                }
-                let arr = builder.finish();
-                ensure_eq!(arr.len(), cardinality);
-
-                Ok(arr.into())
-            }
-            )*
-        }
-    };
-}
-
-read_one_value_array! {
-    { Interval, IntervalArrayBuilder },
-    { Date, DateArrayBuilder },
-    { Time, TimeArrayBuilder },
-    { Timestamp, TimestampArrayBuilder }
-}
-
-fn read_offset(offset_cursor: &mut Cursor<&[u8]>) -> ArrayResult<i64> {
-    match offset_cursor.read_i64::<BigEndian>() {
-        Ok(offset) => Ok(offset),
-        Err(e) => bail!("failed to read i64 from offset buffer: {}", e),
-    }
-}
-
-pub fn read_string_array<B: ArrayBuilder, R: VarSizedValueReader<B>>(
-    array: &PbArray,
-    cardinality: usize,
-) -> ArrayResult<ArrayImpl> {
-    ensure!(
-        array.get_values().len() == 2,
-        "Must have exactly 2 buffers in a string array"
-    );
-    let offset_buff = array.get_values()[0].get_body().as_slice();
-    let data_buf = array.get_values()[1].get_body().as_slice();
-
-    let mut builder = B::new(cardinality);
-    let bitmap: Bitmap = array.get_null_bitmap()?.into();
-    let mut offset_cursor = Cursor::new(offset_buff);
-    let mut data_cursor = Cursor::new(data_buf);
-    let mut prev_offset: i64 = -1;
-
-    let mut buf = Vec::new();
-    for not_null in bitmap.iter() {
-        if not_null {
-            if prev_offset < 0 {
-                prev_offset = read_offset(&mut offset_cursor)?;
-            }
-            let offset = read_offset(&mut offset_cursor)?;
-            let length = (offset - prev_offset) as usize;
-            prev_offset = offset;
-            buf.resize(length, Default::default());
-            data_cursor.read_exact(buf.as_mut_slice()).map_err(|e| {
-                anyhow!(
-                    "failed to read str from data buffer: {} [length={}, offset={}]",
-                    e,
-                    length,
-                    offset
-                )
-            })?;
-            R::read(buf.as_slice(), &mut builder)?;
-        } else {
-            builder.append(None);
-        }
-    }
-    let arr = builder.finish();
-    ensure_eq!(arr.len(), cardinality);
-
-    Ok(arr.into())
-}
