+++ src/common/src/array/mod.rs
@@ -18,7 +18,6 @@ mod arrow;
 mod bool_array;
 pub mod bytes_array;
 mod chrono_array;
-mod column_proto_readers;
 mod data_chunk;
 pub mod data_chunk_iter;
 mod decimal_array;
@@ -29,6 +28,7 @@ mod jsonb_array;
 pub mod list_array;
 mod num256_array;
 mod primitive_array;
+mod proto_reader;
 pub mod stream_chunk;
 mod stream_chunk_iter;
 pub mod stream_record;
@@ -56,7 +56,7 @@ pub use list_array::{ListArray, ListArrayBuilder, ListRef, ListValue};
 pub use num256_array::*;
 use paste::paste;
 pub use primitive_array::{PrimitiveArray, PrimitiveArrayBuilder, PrimitiveArrayItemType};
-use risingwave_pb::data::{PbArray, PbArrayType};
+use risingwave_pb::data::PbArray;
 pub use stream_chunk::{Op, StreamChunk, StreamChunkTestExt};
 pub use struct_array::{StructArray, StructArrayBuilder, StructRef, StructValue};
 pub use utf8_array::*;
@@ -660,43 +660,6 @@ for_all_variants! { impl_array_estimate_size }
 impl ArrayImpl {
     pub fn iter(&self) -> impl DoubleEndedIterator<Item = DatumRef<'_>> + ExactSizeIterator {
         (0..self.len()).map(|i| self.value_at(i))
-    }
-
-    pub fn from_protobuf(array: &PbArray, cardinality: usize) -> ArrayResult<Self> {
-        use self::column_proto_readers::*;
-        use crate::array::value_reader::*;
-        let array = match array.array_type() {
-            PbArrayType::Int16 => read_numeric_array::<i16, I16ValueReader>(array, cardinality)?,
-            PbArrayType::Int32 => read_numeric_array::<i32, I32ValueReader>(array, cardinality)?,
-            PbArrayType::Int64 => read_numeric_array::<i64, I64ValueReader>(array, cardinality)?,
-            PbArrayType::Serial => {
-                read_numeric_array::<Serial, SerialValueReader>(array, cardinality)?
-            }
-            PbArrayType::Float32 => read_numeric_array::<F32, F32ValueReader>(array, cardinality)?,
-            PbArrayType::Float64 => read_numeric_array::<F64, F64ValueReader>(array, cardinality)?,
-            PbArrayType::Bool => read_bool_array(array, cardinality)?,
-            PbArrayType::Utf8 => {
-                read_string_array::<Utf8ArrayBuilder, Utf8ValueReader>(array, cardinality)?
-            }
-            PbArrayType::Decimal => {
-                read_numeric_array::<Decimal, DecimalValueReader>(array, cardinality)?
-            }
-            PbArrayType::Date => read_date_array(array, cardinality)?,
-            PbArrayType::Time => read_time_array(array, cardinality)?,
-            PbArrayType::Timestamp => read_timestamp_array(array, cardinality)?,
-            PbArrayType::Interval => read_interval_array(array, cardinality)?,
-            PbArrayType::Jsonb => {
-                read_string_array::<JsonbArrayBuilder, JsonbValueReader>(array, cardinality)?
-            }
-            PbArrayType::Struct => StructArray::from_protobuf(array)?,
-            PbArrayType::List => ListArray::from_protobuf(array)?,
-            PbArrayType::Unspecified => unreachable!(),
-            PbArrayType::Bytea => {
-                read_string_array::<BytesArrayBuilder, BytesValueReader>(array, cardinality)?
-            }
-            PbArrayType::Int256 => Int256Array::from_protobuf(array, cardinality)?,
-        };
-        Ok(array)
     }
 }
